+++ packages/react/src/react/components/compound/editor/components/redo-undo-buttons.tsx
@@ -3,8 +3,10 @@ import { ArrowClockwiseIcon, ArrowCounterClockwiseIcon } from '@phosphor-icons/r
 import { useCurrentEditor } from '@tiptap/react'
 
 import { Button } from '../../../../../../v2'
-import { BaseIcon } from '../../../primitives/base-icon'
 
+/**
+ * @deprecated
+ */
 export const UndoButton = () => {
   const { editor } = useCurrentEditor()
 
@@ -18,14 +20,16 @@ export const UndoButton = () => {
       onClick={() => {
         editor.chain().focus().undo().run()
       }}
-      aria-label="Undo"
       disabled={!editor.can().undo()}
     >
-      <BaseIcon icon={ArrowCounterClockwiseIcon} weight="regular" />
+      <ArrowCounterClockwiseIcon className="size-8" />
     </Button>
   )
 }
 
+/**
+ * @deprecated
+ */
 export const RedoButton = () => {
   const { editor } = useCurrentEditor()
 
@@ -39,10 +43,9 @@ export const RedoButton = () => {
       onClick={() => {
         editor.chain().focus().redo().run()
       }}
-      aria-label="Redo"
       disabled={!editor.can().redo()}
     >
-      <BaseIcon icon={ArrowClockwiseIcon} weight="regular" />
+      <ArrowClockwiseIcon className="size-8" />
     </Button>
   )
 }