+++ packages/react/v2/components/compound/editor/components/editor-color-picker.tsx
@@ -1,0 +1,64 @@
+'use client'
+import { useState } from 'react'
+import { type Color as ReactAriaColor, parseColor } from 'react-aria-components'
+
+import { SelectionBackgroundIcon } from '@phosphor-icons/react'
+import { useCurrentEditor } from '@tiptap/react'
+
+// TODO: Migrate to v2
+import {
+  ColorPicker,
+  type ColorPickerProps,
+} from '../../../../../src/react/components/primitives/color-picker'
+
+export const EditorColorPicker = ({
+  onPopupClose,
+  ...props
+}: {
+  onPopupClose: (color: ReactAriaColor) => void
+} & ColorPickerProps) => {
+  const [color, setColor] = useState(parseColor('#000000'))
+
+  return (
+    <ColorPicker
+      onPopupOpenChange={(isOpen) => {
+        if (!isOpen) onPopupClose(color)
+      }}
+      // value={parseColor('#AAABBB')} // TODO
+      eyeDropper
+      onChange={setColor}
+      {...props}
+    />
+  )
+}
+
+export const EditorBgColorPicker = () => {
+  const { editor } = useCurrentEditor()
+
+  if (!editor) throw new Error('Editor provider is missing')
+
+  return (
+    <EditorColorPicker
+      onPopupClose={(color) => {
+        editor.chain().setBackColor(color.toString('hex')).run()
+      }}
+      label={<SelectionBackgroundIcon />}
+      buttonClassName="p-4 bg-secondary/25"
+    />
+  )
+}
+
+export const EditorTextColorPicker = () => {
+  const { editor } = useCurrentEditor()
+
+  if (!editor) throw new Error('Editor provider is missing')
+
+  return (
+    <EditorColorPicker
+      onPopupClose={(color) => {
+        editor.chain().setColor(color.toString('hex')).run()
+      }}
+      buttonClassName="p-4 bg-secondary/25"
+    />
+  )
+}