+++ packages/react/v2/components/compound/editor/extensions/custom-link-extension.ts
@@ -1,0 +1,15 @@
+import Link from '@tiptap/extension-link'
+
+export const CustomLinkExtension = Link.extend({
+  inclusive: false,
+}).configure({
+  openOnClick: false,
+  HTMLAttributes: {
+    class: 'custom-link',
+    rel: 'noopener noreferrer',
+    target: '_blank',
+  },
+  protocols: ['http', 'https', 'mailto', 'tel'],
+  autolink: true,
+  linkOnPaste: true,
+})