+++ packages/react/v2/components/compound/editor/extensions/index.ts
@@ -1,0 +1,3 @@
+export * from './background-color-extension'
+export * from './custom-link-extension'
+export * from './selection-extension'