+++ packages/rest/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @genseki/rest
 
+## 0.1.0-alpha.77
+
+## 0.1.0-alpha.76
+
+## 0.1.0-alpha.75
+
+## 0.1.0-alpha.74
+
 ## 0.1.0-alpha.73
 
 ## 0.1.0-alpha.72